feat: create a standalone tags dialog
