+++ src/containers/MyNdla/Folders/components/ResourceWithMenu.tsx
@@ -6,13 +6,14 @@
  *
  */
 
-import { DeleteBinLine, FolderLine, LinkMedium } from "@ndla/icons";
+import { DeleteBinLine, FolderLine, HashTag, LinkMedium } from "@ndla/icons";
 import { useCallback, useContext, useMemo, useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { AuthContext } from "../../../../components/AuthenticationContext";
 import { AddResourceToFolderModalContent } from "../../../../components/MyNdla/AddResourceToFolderModalContent";
 import { DeleteModalContent } from "../../../../components/MyNdla/DeleteModalContent";
 import { ListResource } from "../../../../components/MyNdla/ListResource";
+import { UpdateResourceTagsDialogContent } from "../../../../components/MyNdla/UpdateResourceTagsDialogContent";
 import { useToast } from "../../../../components/ToastContext";
 import config from "../../../../config";
 import { GQLFolder, GQLMyNdlaResource, GQLMyNdlaResourceMetaFragment } from "../../../../graphqlTypes";
@@ -105,6 +106,14 @@ export const ResourceWithMenu = ({ resource, loading, resourceMeta, selectedFold
           />
         ),
         variant: "destructive",
+      },
+      {
+        type: "dialog",
+        value: "updateTags",
+        icon: <HashTag />,
+        text: t("myNdla.resource.updateTags"),
+        isModal: true,
+        modalContent: (close) => <UpdateResourceTagsDialogContent onClose={close} resource={resource} />,
       },
     ];
   }, [examLock, onDeleteResource, resource, selectedFolder, t, toast]);
