+++ src/messages/messagesNB.ts
@@ -162,6 +162,7 @@ const messages = {
       titleUpdated: "Tittel oppdatert",
       tagsUpdated: "Emneknagger oppdatert",
       tagsUpdatedFailed: "Klarte ikke å oppdatere emneknagger",
+      updateTags: "Oppdater emneknagger",
       show: "Vis",
       save: "Lagre ressurs",
     },
